webapi/tct-application-tizen-tests/application/ApplicationManager_launchAppControl_with_errorCallback.html

<!DOCTYPE html>

<!--

Copyright (c) 2013 Samsung Electronics Co., Ltd.

Licensed under the Apache License, Version 2.0 (the License);
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.



Authors:
        Mariusz Polasinski <m.polasinski@samsung.com>


-->

<html>

<head>
<title>ApplicationManager_launchAppControl_with_errorCallback</title>
<meta charset="utf-8">
<script type="text/javascript" src="../resources/testharness.js"></script>
<script type="text/javascript" src="../resources/testharnessreport.js"></script>
<script type="text/javascript" src="support/app_common.js"></script>
</head>

<body>
<div id="log"></div>
<script type="text/javascript">

//==== TEST: ApplicationManager_launchAppControl_with_errorCallback
//==== LABEL Check using launchAppControl method (with optional argument errorCallback) in ApplicationManager interface
//==== SPEC Tizen Web API:Application:Application:ApplicationManager:launchAppControl M
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/application.html
//==== PRIORITY P1
//==== ONLOAD_DELAY 30
//==== TEST_CRITERIA MERRCB

setup({timeout: 30000});

var t = async_test(document.title, {timeout: 30000}),
    launchSuccess, launchError, appcontrol;

setup_launch(t, TCT_APPCONTROL_APPID, function () {
    launchSuccess = t.step_func(function () {
        assert_unreached("this function should not be used");
    });

    launchError = t.step_func(function () {
        assert_not_launch(t, TCT_APPCONTROL_APPID, function () {
            t.done();
        });
    });

    appcontrol = new tizen.ApplicationControl(
        TCT_APPCONTROL_LAUNCH_APPCONTROL_OPERATION,
        TCT_APPCONTROL_LAUNCH_APPCONTROL_URI,
        TCT_APPCONTROL_LAUNCH_APPCONTROL_MIME_INVALID
    );

    tizen.application.launchAppControl(appcontrol, null, launchSuccess, launchError);
});

</script>
</body>
</html>
